feat: integrate unified content rendering system
- Integrate [slug].vue to use centralized contentUtils.js rendering functions
- Remove duplicate renderTipTapNodes and renderTipTapMarks functions (260+ lines)
- Clean up unused imports (isExternalVideoUrl, getEmbedUrl) from [slug].vue
- Fix variable redeclaration errors in contentUtils.js (text vs nodeText)
- Add block scopes to all switch case statements to prevent lexical declaration conflicts
- Ensure consistent content rendering across announcement and meme systems
- Reduce code duplication and improve maintainability of content processing logic
