+++ src/pages/memes/detail/[slug].vue
@@ -724,9 +724,9 @@ import recommendationService from '@/services/recommendationService'
 
 // 工具函數
 import { getId, getMemeId, getMemeSlug } from '@/utils/dataUtils'
-import { isExternalVideoUrl, getEmbedUrl } from '@/utils/mediaUtils'
 import { getShareOptions, handlePlatformShare } from '@/utils/shareUtils'
 import { requireLogin } from '@/utils/authUtils'
+import { renderContentToHtml } from '@/utils/contentUtils'
 
 // 路由和狀態
 const route = useRoute()
@@ -916,230 +916,12 @@ const sources = computed(() => {
     : []
 })
 
-// 渲染 TipTap JSON 內容
+// 渲染 TipTap JSON 內容（使用共用工具函數）
 const renderTipTapContent = (content) => {
   if (!content) return ''
 
-  // 如果是字串，可能是舊的 markdown 格式，使用原本的處理方式
-  if (typeof content === 'string') {
-    return content
-      .replace(/\*\*(.*?)\*\*/g, '<strong>$1</strong>') // 粗體
-      .replace(/\*(.*?)\*/g, '<em>$1</em>') // 斜體
-      .replace(
-        /`(.*?)`/g,
-        '<code class="bg-surface-100 px-1 rounded">$1</code>',
-      ) // 行內代碼
-      .replace(
-        /^### (.*$)/gim,
-        '<h3 class="text-lg font-bold mt-4 mb-2">$1</h3>',
-      ) // 三級標題
-      .replace(
-        /^## (.*$)/gim,
-        '<h2 class="text-xl font-bold mt-6 mb-3">$1</h2>',
-      ) // 二級標題
-      .replace(
-        /^# (.*$)/gim,
-        '<h1 class="text-2xl font-bold mt-8 mb-4">$1</h1>',
-      ) // 一級標題
-      .replace(/^- (.*$)/gim, '<li class="ml-4">$1</li>') // 無序列表
-      .replace(/^\d+\. (.*$)/gim, '<li class="ml-4">$1</li>') // 有序列表
-      .replace(/\n\n/g, '</p><p class="mt-4">') // 段落
-      .replace(/^(.+)$/gm, '<p class="leading-relaxed">$1</p>') // 一般文字
-  }
-
-  // 如果是 TipTap JSON 格式
-  if (
-    typeof content === 'object' &&
-    content.type === 'doc' &&
-    content.content
-  ) {
-    return renderTipTapNodes(content.content)
-  }
-
-  return ''
-}
-
-// 遞迴渲染 TipTap 節點
-const renderTipTapNodes = (nodes) => {
-  if (!Array.isArray(nodes)) return ''
-
-  return nodes
-    .map((node) => {
-      switch (node.type) {
-        case 'paragraph':
-          return `<p>${renderTipTapMarks(node.content)}</p>`
-
-        case 'heading': {
-          const level = node.attrs?.level || 1
-          return `<h${level}>${renderTipTapMarks(node.content)}</h${level}>`
-        }
-
-        case 'image': {
-          const {
-            src,
-            alt,
-            annotation,
-            size = 'm',
-            orientation = 'landscape',
-          } = node.attrs || {}
-
-          // 根據尺寸和方向設定樣式
-          const getSizeStyles = (size, orientation) => {
-            const landscapeMap = {
-              s: { width: '320px', maxWidth: '320px' },
-              m: { width: '640px', maxWidth: '640px' },
-              l: { width: '960px', maxWidth: '960px' },
-              full: { width: '100%', maxWidth: '100%' },
-            }
-            const portraitMap = {
-              s: { width: '240px', maxWidth: '240px' },
-              m: { width: '480px', maxWidth: '480px' },
-              l: { width: '720px', maxWidth: '720px' },
-              full: { width: '100%', maxWidth: '100%' },
-            }
-            const sizeMap =
-              orientation === 'portrait' ? portraitMap : landscapeMap
-            return sizeMap[size] || sizeMap.m
-          }
-
-          const sizeStyles = getSizeStyles(size, orientation)
-          const imageStyle =
-            size === 'full'
-              ? 'width: 100%; max-width: 100%; height: auto; display: block;'
-              : `width: 100%; max-width: ${sizeStyles.maxWidth}; height: auto; display: block;`
-
-          let imageHtml = `<div class="custom-image-wrapper" style="width: 100%; max-width: ${sizeStyles.maxWidth};"><img src="${src}" alt="${alt || ''}" class="custom-image" style="${imageStyle}" />`
-          if (annotation) {
-            imageHtml += `<p class="image-annotation">${annotation}</p>`
-          }
-          imageHtml += `</div>`
-          return imageHtml
-        }
-
-        case 'videoEmbed': {
-          const {
-            src: videoSrc,
-            size = 'm',
-            orientation = 'landscape',
-            annotation,
-          } = node.attrs || {}
-
-          if (videoSrc && isExternalVideoUrl(videoSrc)) {
-            // 根據尺寸和方向設定樣式
-            const getSizeStyles = (size, orientation) => {
-              const landscapeMap = {
-                s: { width: '100%', maxWidth: '480px' },
-                m: { width: '100%', maxWidth: '640px' },
-                l: { width: '100%', maxWidth: '960px' },
-                full: { width: '100%', maxWidth: 'none' },
-              }
-              const portraitMap = {
-                s: { width: '100%', maxWidth: '360px' },
-                m: { width: '100%', maxWidth: '450px' },
-                l: { width: '100%', maxWidth: '540px' },
-                full: { width: '100%', maxWidth: 'none' },
-              }
-              const sizeMap =
-                orientation === 'portrait' ? portraitMap : landscapeMap
-              return sizeMap[size] || sizeMap.m
-            }
-
-            const sizeStyles = getSizeStyles(size, orientation)
-            const aspectRatio = orientation === 'portrait' ? '9/16' : '16/9'
-
-            let videoHtml = `<div class="video-embed-wrapper" style="width: 100%; max-width: ${sizeStyles.maxWidth};">
-            <div class="video-embed-container">
-            <iframe src="${getEmbedUrl(videoSrc)}"
-                    class="video-embed-iframe"
-                    frameborder="0"
-                    allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-                    allowfullscreen
-                    style="aspect-ratio: ${aspectRatio};">
-            </iframe>
-          </div>`
-
-            if (annotation) {
-              videoHtml += `<p class="video-annotation">${annotation}</p>`
-            }
-
-            videoHtml += `</div>`
-            return videoHtml
-          }
-          return ''
-        }
-
-        case 'bulletList':
-          return `<ul>${renderTipTapNodes(node.content)}</ul>`
-
-        case 'orderedList':
-          return `<ol>${renderTipTapNodes(node.content)}</ol>`
-
-        case 'listItem':
-          return `<li>${renderTipTapMarks(node.content)}</li>`
-
-        case 'blockquote':
-          return `<blockquote >${renderTipTapMarks(node.content)}</blockquote>`
-
-        case 'codeBlock':
-          return `<pre><code>${node.content?.[0]?.text || ''}</code></pre>`
-
-        case 'horizontalRule':
-          return `<hr />`
-
-        default:
-          return ''
-      }
-    })
-    .join('')
-}
-
-// 渲染 TipTap 標記
-const renderTipTapMarks = (content) => {
-  if (!Array.isArray(content)) return ''
-
-  return content
-    .map((item) => {
-      if (item.type === 'text') {
-        let text = item.text || ''
-
-        // 處理標記
-        if (item.marks) {
-          item.marks.forEach((mark) => {
-            switch (mark.type) {
-              case 'bold':
-                text = `<strong>${text}</strong>`
-                break
-              case 'italic':
-                text = `<em>${text}</em>`
-                break
-              case 'underline':
-                text = `<u>${text}</u>`
-                break
-              case 'strike':
-                text = `<del>${text}</del>`
-                break
-              case 'code':
-                text = `<code class="bg-surface-100 text-surface-800 p-1 rounded-md dark:bg-surface-900 dark:text-surface-300 font-mono">${text}</code>`
-                break
-              case 'link': {
-                const { href, target, rel } = mark.attrs || {}
-                text = `<a href="${href}" target="${target || '_blank'}" rel="${rel || 'noopener noreferrer'}" class="decoration-0 cursor-pointer font-medium text-surface-800 transition-all p-1 dark:text-surface-300 hover:text-primary-500 underline">${text}</a>`
-                break
-              }
-            }
-          })
-        }
-
-        return text
-      }
-
-      if (item.type === 'hardBreak') {
-        return '<br />'
-      }
-
-      return ''
-    })
-    .join('')
+  // 使用 contentUtils.js 中的統一渲染函數
+  return renderContentToHtml(content, 'json')
 }
 
 // 處理 API 響應數據
